imphook: Update for new hook signature, take "module name" param.
The idea was to avoid parsing "superfluous" params into hook functions,
and indeed, it's possible to pass only filename path as long as we
*don't* deal with importing of Python source modules. We just populate
module, and can set its __name__ afterwards.

But as soon as we deal with Python modules, we need a module name, and
we need to set it before executing module top-level code, because this
top-level code may access/check it.

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>