mod_funkw_naive: "Naive" version of "function" -> "lambda" .py hook.
This import hook allows to use keyword "function" instead of "lambda",
e.g. "f = function x: x + 1".

The implementation does naive substring replacement over entire module
source code. It's brought in as an example of how to *not* write a
Python source transformation hook. (Going to be superseded with proper
hook working on the token stream.)

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>